Split GitlabInstanceInterface into interfaces
